backend/requirements.txt
fastapi==0.135.1
uvicorn[standard]==0.34.0
langgraph==1.1.0
langchain-anthropic==1.3.4
tavily-python==0.7.22
langsmith==0.7.16
aiosqlite==0.20.0
python-dotenv==1.1.0
pydantic==2.11.0
